refactor facet runtime APIs and isolate adapter extension surface
