+++ crates/dspy-rs/src/predictors/predict.rs
@@ -9,11 +9,10 @@ use tracing::{debug, trace};
 
 use crate::adapter::Adapter;
 use crate::bamltype::baml_types::BamlMap;
-use crate::bamltype::compat::{BamlValueConvert, ToBamlValue};
 use crate::core::{FieldSpec, MetaSignature, Module, Optimizable, Signature};
 use crate::{
-    BamlValue, CallResult, Chat, ChatAdapter, Example, GLOBAL_SETTINGS, LM, LmError, LmUsage,
-    PredictError, Prediction,
+    BamlValue, BamlValueConvert, CallResult, Chat, ChatAdapter, Example, GLOBAL_SETTINGS, LM,
+    LmError, LmUsage, PredictError, Prediction, ToBamlValue,
 };
 
 pub struct Predict<S: Signature> {
